+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 - [#131](https://github.com/os2display/display-client/pull/131)
   - Moved localstorage calls into helper class AppStorage.
   - Cleaned up project structure.
+  - Fixed async flow for release check.
+  - Added previous boot timestamp (pb) to url.
 - [#130](https://github.com/os2display/display-client/pull/130)
   - Fixed issue with svg loading.
 - [#127](https://github.com/os2display/display-client/pull/127)